fix: cast explícito no UPDATE de remover_item evita AmbiguousParameterError
asyncpg não conseguia inferir o tipo de $1 usado como int (quantidade)
e como multiplicador de decimal (valor_total). Cast $1::int e $1::numeric resolve.

+++ app/database.py
@@ -270,7 +270,7 @@ async def remover_item(comanda_id: UUID, produto_nome: str, quantidade: int) -> 
                 await conn.execute(
                     """
                     UPDATE itens_comanda
-                    SET quantidade = $1, valor_total = $1 * valor_unitario
+                    SET quantidade = $1::int, valor_total = $1::numeric * valor_unitario
                     WHERE id = $2
                     """,
                     nova_qtd,
